Added the tube part converter - no more disappearing tubes! :D
